comments added to the timemanager script
why do we use this script, which has no other function but keep track of time to initiate create and destroy news objects? basically everything that has anything to do with the newsfeed logic is moved into the news script to ease debugging and maintenance in the future. I believe that the sole reason of this scripts' existence should remain with time related issues